requirements.txt
boto3==1.41.5
botocore==1.41.5
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
numpy==2.3.5
pandas==2.3.3
pyarrow==22.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
